+++ java/src/main/java/com/github/cowwoc/requirements/java/type/StringValidator.java
@@ -87,6 +87,16 @@ public interface StringValidator extends
 	StringValidator doesNotContain(String unwanted);
 
 	/**
+	 * Ensures that the value matches a regular expression.
+	 *
+	 * @param regex the regular expression
+	 * @return this
+	 * @throws NullPointerException     if the value is null
+	 * @throws IllegalArgumentException if the value does not match {@code regex}
+	 */
+	StringValidator matches(String regex);
+
+	/**
 	 * Ensures that the value is empty.
 	 *
 	 * @return this