+++ docs/Changelog.md
@@ -31,6 +31,7 @@ See https://github.com/cowwoc/requirements.java/commits/master for a full list.
         design for this yet.
     16. Added `ObjectValidator.isX()` methods to downcast to known types.
     17. Replaced thread context with `Validators.getContext()`, `putContext()`, `removeContext()`.
+    18. Added `StringValidator.matches(regex)`.
 * Bugfixes:
     * `StringValidator/Verifier.asShort()`, `asInteger()` and `asLong()` were not handling the case where a
       string could not be converted to a number.